feat: Refactor user management functions and enhance user creation and status update logic
